Remove double count problem from metric_handler.
Previously it was possible for a user to be double counted in the
metrics, first as an anonymous user and then as a logged in user
once they completed login.

volatile_hash has been extended to allow key, value pairs to be
deleted. This allows the metric_handler to remove session_id's
from cache's, removing the possibility of double counting.