Message Processor publishes to SNS
So that this commit does not break the dashboard, metrics and events
are still sent to it via a http POST. SNS notifications are also
published. This will allow changes to be made to the dashboard to listen
to the SNS notifications. When this is complete, the http POST can be
removed, giving a cleaner decoupling between this metric_handler process
and the dashboard.

Other interested processes can subscribe to SNS notifications once this
is live.